ch-loader/requirements.txt
mysql-connector-python
clickhouse-connect
pandas


